requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-writable-nested==0.7.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
